Remove unused compressedRefs anchors in dead trees
Dead trees elimination removes indirect loads under compressedRefs
anchors when they have reference count 1, i.e. when their results are
unused. But when the result of such a load is used anywhere, its anchor
is left in place. So far this determination has been made only when the
pass encounters the anchor, which means that it counts later uses even
if those uses are themselves about to be removed.

Dead trees elimination now revisits compressedRefs anchors for loads at
the end of the pass, in order to determine whether the anchored loads
have become unused, and remove any that have.

Signed-off-by: Devin Papineau <devinmp@ca.ibm.com>