Upgrade lint job environment on Travis CI to Trusty
Currently, the OMRChecker lint job on Travis CI runs
in Precise. It is in the process of being decommisoned,
and the tentative official end of support for Precise
by Travis CI was April 1st 2018. This patch upgrades the
environment to Trusty, and also makes use of clang-3.8
instead of clang-3.4.

Signed-off-by: Nazim Uddin Bhuiyan <nazimudd@ualberta.ca>